fix: customize components of movies screen
