--- src/components/movies/index.ts
@@ -1,3 +1,0 @@
-import Movies from './Movies';
-
-export default Movies;
